Supports macos screen capture and optimizes the connection process. (#59)

+++ common/src/frame.rs
@@ -53,6 +53,8 @@
 
 use std::{ffi::c_void, ptr::null};
 
+use serde_repr::{Deserialize_repr, Serialize_repr};
+
 /// A sample from the audio stream.
 #[repr(C)]
 #[derive(Debug)]
@@ -78,8 +80,8 @@ impl Default for AudioFrame {
 }
 
 /// Video frame format.
-#[repr(C)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[repr(i32)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize_repr, Serialize_repr)]
 pub enum VideoFormat {
     BGRA,
     RGBA,
@@ -88,8 +90,8 @@ pub enum VideoFormat {
 }
 
 /// Subtype of the video frame.
-#[repr(C)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[repr(i32)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize_repr, Serialize_repr)]
 pub enum VideoSubFormat {
     /// This video frame is from Core video, a type exclusive to the Macos
     /// platform.
@@ -113,7 +115,7 @@ pub struct VideoFrame {
     /// format and the data corresponds to the plane of the corresponding
     /// format, All other sub formats use `data[0]`.
     pub data: [*const c_void; 3],
-    pub linesize: [usize; 3],
+    pub linesize: [u32; 3],
 }
 
 unsafe impl Sync for VideoFrame {}
